+++ RELEASENOTES.md
@@ -82,6 +82,8 @@
     *   Avoid potential delays caused by handling routing change callbacks at
         the beginning of playback
         ([#2646]()https://github.com/androidx/media/issues/2646).
+    *   Allow codec reuse for EAC3, EAC3-JOC and AC-4 formats
+        ([#1346](https://github.com/androidx/media/pull/1346)).
 *   Video:
 *   Text:
 *   Metadata: